Added Scripts for Saving and Loading
Added the SavableObject Component, which marks a GameObject to be saved.
Added the SavableList Component which saves, loads, and destroys SavableObjects that are children of it.
The Constants.cs file holds the paths to access objects that are to be loaded, and can be used to store future static data.