Added Hazards UI
Added and modified scripts on the Level Manager object to collect Hazard scripts found in loaded objects. Hazard scripts should be added to the prefab of every object that will be considered a hazard for scoring purposes. Also added another panel to the Pause Menu that displays the number of hazards to be dealt with.